+++ hudi-sync/hudi-hive-sync/src/main/java/org/apache/hudi/hive/HiveStylePartitionValueExtractor.java
@@ -18,7 +18,7 @@
 
 package org.apache.hudi.hive;
 
-import org.apache.hudi.sync.common.model.partextractor.PartitionValueExtractor;
+import org.apache.hudi.sync.common.model.PartitionValueExtractor;
 
 import java.util.Collections;
 import java.util.List;
