+++ hudi-sync/hudi-hive-sync/src/test/java/org/apache/hudi/hive/TestPartitionValueExtractor.java
@@ -18,8 +18,6 @@
 
 package org.apache.hudi.hive;
 
-import org.apache.hudi.sync.common.model.partextractor.SlashEncodedHourPartitionValueExtractor;
-
 import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.List;
