+++ hudi-sync/hudi-sync-common/src/main/java/org/apache/hudi/sync/common/model/PartitionValueExtractor.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.hudi.sync.common.model.partextractor;
+package org.apache.hudi.sync.common.model;
 
 import java.io.Serializable;
 import java.util.List;
